rename rustwrapper folder
