+++ src/lib.rs
@@ -1,3 +1,3 @@
 // Distributed under the MIT software license, see the accompanying
 // file LICENSE or http://www.opensource.org/licenses/mit-license.php.
-pub mod rustwrapper;
+pub mod rust_wrapper;
