+++ tests/rust/dilithium/dilithium_test.rs
@@ -1,7 +1,7 @@
 use pqcrypto_traits::sign::SignedMessage;
 // Distributed under the MIT software license, see the accompanying
 // file LICENSE or http://www.opensource.org/licenses/mit-license.php.
-use qrllib::rustwrapper::dilithium::dilithium::Dilithium;
+use qrllib::rust_wrapper::dilithium::dilithium::Dilithium;
 
 #[test]
 fn sign_keypair() {
